[AS-606] adjust Wallets service structure
