+++ src/services/index.ts
@@ -4,7 +4,7 @@
 
 export * from "../application/services/Fee";
 export * from "../infrastructure/adapters/Crypto";
-export * from "./Wallets";
+export * from "../application/services/Wallets";
 export * from "../application/services/Resubmit";    
 export * from "../infrastructure/adapters/Signer";
 export * from "../application/services/AssetsService";
@@ -12,7 +12,7 @@ export * from "../application/services/Fee";
 export * from "../application/services/TxHistory";
 
 export { default as BinaryWriter } from "../infrastructure/adapters/BinaryWriter";
-export { default as WalletsService } from "./Wallets";
+export { default as WalletsService } from "../application/services/Wallets";
 export { default as CryptoService } from "../infrastructure/adapters/Crypto";
